feat: Deprecate old fruit based tokens and map them to new color palette value (#159)
Fixes: https://github.com/Workday/canvas-tokens/issues/158
Deprecate old fruit based color tokens and map them to new values based on the new color palette.

Components

Release Note:
We've deprecated the fruit based color tokens and provided a mapping to new base tokens for the new color palette. In the docs, we provide a mapping of the new CSS token and JS token.

Co-authored-by: manuel.carrera <manuel.carrera@workday.com>
Co-authored-by: @alanbsmith <a.bax.smith@gmail.com>
Co-authored-by: @RayRedGoose <48605821+RayRedGoose@users.noreply.github.com>